deduplicate_works: move functions that don't manipulate the state variables to the helpers module
